refactor: 공통으로 사용하는 toInt 함수 utils에 선언하여 재사용
