+++ src/mocks/handlers/lostItems.handlers.ts
@@ -1,13 +1,6 @@
 import { http, HttpResponse } from 'msw';
 import { getSummary, getDetail, getEtcDetail } from '../selectors/lostItems.selectors';
-
-function toInt(value: unknown): number | undefined {
-  const v = Array.isArray(value) ? value[0] : value;
-  if (v == null) return undefined;
-  if (typeof v !== 'string') return undefined;
-  const n = Number(v);
-  return Number.isFinite(n) ? n : undefined;
-}
+import { toInt } from '../utils/toInt';
 
 export const lostItemsHandlers = [
   http.get('/api/lost-items/summary', ({ request }) => {
